ocrFunction/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.18.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1
requests==2.31.0